+++ example/Get-HtmlTemplate.ps1
@@ -1,6 +1,6 @@
 <#
 .SYNOPSIS
-	Renders a HTML document from the given template.
+	Renders a HTML document from the given PowerShell script.
 #>
 Import-Module "$PSScriptRoot/../Html.psd1" # TODO Import-Module Html
 $html = & "$PSScriptRoot/../res/Document.ps1"