+++ example/Write-HtmlTemplate.ps1
@@ -1,8 +1,7 @@
 <#
 .SYNOPSIS
-	Renders a HTML document from the given template.
+	Renders a HTML document from the given PowerShell template file.
 #>
 Import-Module "$PSScriptRoot/../Html.psd1" # TODO Import-Module Html
-
-$html = & "$PSScriptRoot/../res/Document.ps1"
-Write-Host $html
+$html = Write-HtmlTemplate "$PSScriptRoot/../res/Document.ps1"
+Write-Output $html